Bug 265084 Need to deal with integer overflows in reflow r=bz sr=bz

+++ gfx/public/nsCoord.h
@@ -150,7 +150,7 @@ inline float NSCoordToFloat(nscoord aCoord) {
 #ifdef NS_COORD_IS_FLOAT
   NS_ASSERTION(!NS_IEEEIsNan(aCoord), "NaN encountered in float conversion");
 #endif
-  return aCoord;
+  return (float)aCoord;
 }
 
 #endif /* NSCOORD_H */
